oiiotool --echo
This adds an --echo MESSAGE command to oiiotool, which will print the
message to the console at that point the left-to-right evaluation of
command line arguments.

For example:

    oiiotool test.tif --resize 256x0 --echo "result is {TOP.width}x{TOP.height}"

This will resize the input to be 256 pixels wide and automatically size
it vertically to preserve the original aspect ratio, and then print
a message to the console revealing the resolution of the resulting image.