Add some extra checks to exif parser to avoid memory access issues
This is an attempt to solve #1635. This avoids the crash by ignoring
any directory entry which is outside of the parsing buffer (that would
be really weird if there is a legit memory pointer to an outside memory
block).

There are much more places where there is potential crash is burried
in here, but those i don't feel like going into.

For such things it's really good idea to trust no input data and verify
it, otherwise it's really easy to run into crashes and even possibly
security issues.

Maybe it's also not a bad idea to add an option tyo ignore exif headers
reading for the applications which doesn't need exif.