JPEG: auto conform to supported channel count. (#1643)
JFIF files, as you know, only support 1 or 3 channels (grayscale and
RGB, respectively).

Previously, we accepted the special case of 4 channel output (presumably
RGBA) by silently dropping the last channel (A, not representable in
JFIF files), but for any other channel count (2, or >= 5) would just
allow open() to fail.

With this patch, we make the JPEG writer even more forgiving. For
nchannels > 3, we truncate to just the first three channels, and for 2
channels, we truncate to just the first channel (outputting grayscale).

The reason why we want this extra flexibility is that JPEG is such a
common format for "proxies", "thumbnails", and other kinds of embedded
or web-facing images, where we might want to just quickly generate such
a proxy for arbitrary images without wanting to think hard about the
logic of how to conform unusual channel counts to JPEG's limitations.
And let's face it, if the user really wanted high fidelity, they would
not be using JPEG -- a usually-lossy, 8-bit-only, 1-and-3-channel-only
format -- and so silently dropping a channel that can't be represented
is probably not the biggest compromise being made to the original data
when saving to this format.