Text drawing overhaul (#1646)
1. Extend ImageBufAlgo::render_text in several ways:
(a) Text color is now passed as an array_view<const float>, which is
    safer than passing a raw pointer and hoping it points to enough
    data.
(b) New parameters: alignx and aligny, which can be used to control
    the alignment of the text with the (x,y) coordinates passed.
    Options for alignx are ImageBufAlgo::TextAlignX::Left (default),
    Right, and Center, and options for aligny are TextAlignY::Baseline
    (default), Top, Bottom, Center.
(c) New parameter: int shadow, default to 0 (old behavior), with nonzero
    values causing the rendered text to be dilated with black halo to be
    more visible on backgrounds with similar color.
(d) The destination image no longer is required to be initialized. If not,
    it will be initialized to be just large enough to hold the rendered
    text. Pre-initialized images will continue to draw the text overtop
    of the existing pixel values (as if an "over" composite).

2. Add a new ImageBufAlgo::text_size() function, which is related to
IBA::render_text(), except that rather than modifying an ImageBuf, it
merely computes the bounding box (relative to the left edge of the
baseline of the first character) as a ROI. But with this knowledge, it
becomes easy to do things like align or center text in tricky ways.

3. oiiotool --text has also been extended to have optional xalign= and
yalign= modifiers to align the text to the coordinates
passed. Horizontal alignment choices are "left" (default), "right", and
"center", and vertical alignment choices are "baseline" (default),
"top", and "bottom." Also, an optional modifier shadow= can control the
drop shadow.

So, for example, to make your text centered along the horizontal axis of
an image at scanline 100, and use some shadowing to clarify the text,
you could:

    oiiotool in.exr -text:x={TOP.width/2}:y=100:xalign=center:shadow=1 \
        "Hello, world" -o out.exr