Improve TIFF I/O of color separation images (1658)
I had been assuming that for TIFF photometric mode "separate",
it would always be 4 channels and they would be CMYK. Not so!

Turns out that there's a tag called InkSet which says if it's really
CMYK or not. If it's not, check the InkNames to provide channel names.
Also, the auto-conversion of CMYK to RGB should only happen if it's
truly CMYK, because the conversion wouldn't be correct for arbitrary inks.

On output, also, if the request is made to output color separated mode
(by setting the attribute "tiff:ColorSpace" to "CMYK" or "color
separated"), set those attribute as correctly as we can -- to true CMYK
if it appears to be so (4 channels, named C,M,Y,K or Cyan, Magenta,
Yellow, Black), but otherwise say it's custom inks and output the
channel names as ink names.

Aside: It has always bothered us that TIFF doesn't support channel names
(like OpenEXR does). But it does support the InkNames tag. What if we
always output the channel names as InkNames, even for non-separated
photometric mode. Would there be any negative consequences? Would apps
ever misbehave in the presence of an InkNames tag for an RGB photometric
image, or would it just be ignored?