atomic.h improvements (1661)
1. atomic_min/max: Some web reading makes me think that our
implementation of atomic_min and atomic_max ought to use
compare_exchange_weak rather than compare_exchange_strong. As I
understand it, the stricter requirements of the strong versions
(basically, guaranteeing to never have a "false negative" where it
thinks it needs to try again but it's actually unnecessary) can on some
architectures lead to much worse performance.

2. New atomic_fetch_add for std::atomic<float> and <double>. They
exist, but don't in the current standard support atomic add (so we
implement them in terms of compare_exchange, which is supported).