function_view (1660)
New file, function_view.h, contains the definition for a function_view<>
template that is a lightweight, non-owning wrapper for a callable
object. It has much less overhead than a std::function<>, but at the
expense of being non-owning and needing to ensure that its use does not
outlast the lifetime of the callable entity it wraps (very analogous to
the limitations of string_view and array_view).