oiiotool -resize/-resample: revise logic for nonzero origin, crop, overscan
The 'size' specification argument to oiiotool --resize and --resample
had very unintiutive behavior for images where the display and pixel
data windows were not the same, especially if the data window had a
nonzero origin (such as crop or overscan cases).

The underlying ImageBufAlgo::resize is mapping a display window and
is more straightforward. So in this patch, we take the oiiotool --resize
(and --resample) and interpret it as the desired display window size,
and then compute the approprate pixel window from it, rather than the
other way around (which was really strange).

So now you are saying: "oiiotool, here is the size and offset of the
display window I want for the output. Resize the input image so that its
original display window maps exactly to the new display window I've told
you about. Figure out what pixel data window you will need in the output
to ensure that all of the data pixels in the input will be represented."

For the usual 99% case where display == data and origins are 0, the
behavior does not change. But this is definitely easier to control for
the oddball cases.