OSX rpath fix and Makefile INSTALL_PREFIX setting (1671)
1. Some complaints (including failures in OSL TravisCI builds) led me to
believe that the 'cmake_policy (SET CMP0042 OLD)' was not correct and
that we should remove that and set MACOS_RPATH ON. This seems to clean
up some rpath troubles on OSX. It should not have any effect on non-OSX
builds.

2. For the Makefile wrapper, allow INSTALL_PREFIX to be set, which is
passed along to cmake as CMAKE_INSTALL_PREFIX. It defaults to the usual
of ./dist/PLATFORM, but this lets you override to, for example, instruct
the build to put the full installed files in another location. You could
always do this if invoking cmake directly, but it was tricky for those
who preferred to use the make wrapper.