Refactor ParamValueList
We can remove most of the complexity of this class by making
ParamValueList inherit from std::vector<ParamValue> and just add the few
extra methods we want (like a version of find that is case-insensitive
on name, but excludes potential matches that are the wrong data type).
In other words, a change of "has-a" to "is-a" makes most of the code go
away.