Overhaul handling of gamma-corrected images. (1684)
Some (older, not very color-smart) image formats store a gamma exponent
-- neither defining a good color space nor being in a standard like
sRGB.  We were previously haphazardly labeling these as color space
"GammaCorrected", storing the gamma value in the "oiio:Gamma" metadata,
and hoping for the best. But the best, unfortunately, was total
ignorance by our standard color conversion functions (including the
likes of `oiiotool --autocc`).  And we just have too many places where
simply the name of the color space itself (without access to other
metadata) is assumed to be enough to perform color conversions.

So this patch does the following:

1. Adopts a new convention of calling the color space of such images
"GammaCorrectedX.Y", where the X.Y is the decimal gamma value. As an
example, we might now set the "oiio:ColorSpace" metadata for such files
to "GammaCorrected2.2".

2. Follow this convention when reading and writing the formats
supporting an explicit gamma exponent, including Cineon, DPX, HDR/RGBE,
PNG, RLA, Targa.

3. Augment the color transformation engine to recognize these names,
decode the exponent from the name, and thus be capable of (at the very
least) transforming between linear and gamma space. This means that
`oiiotool --autocc` should now be able to properly linearize input files
of this kind (previously it would give a warning and fail to linearize).
And also, it means that you can apply gamma corrections using commands
such as `oiiotool in.exr -colorconvert linear GammaCorrected2.2 -o out.png`

4. Makes `iv` recognize plain old gamma corrected images and not impose
a second correction on it. This is a band-aid, and what it really needs
is a full-blown OCIO support. But it's still a small step forward in
seeing some images properly that previously were incorrect.