deep_holdout
The deep holdout function is one that culls the deep samples in one deep
image (A) that are behind the opaque depth of a second deep image (B).
No samples of B are included in the output, it is only used to mask
parts of A. Samples of A that straddle the opaque frontier are split.

This is implemented as ImageBufAlgo::deep_holdout(), and also in
oiiotool --deepholeout.