IBA::deep_holdout -- fix bug when channels didn't match
deep_holdout() was crashing when the holdout image had more channels
than the input image. It was improper to rely on IBAPrep's default behavior
of setting the result channels to the maximum of the channels in any of
the input images. In this case, we want to force it to match the "shape"
of just the first (source) input, not the second input that only serves
to act as the depth holdout.

While I was there, though, I tightened up the logic of how IBAPrep deals
with the force_spec parameter, and also made that parameter a const ptr,
as it always should have been.