bool[48]::setcomp() is a safer version of int& operator[]
The int& operator[]() was always sketchy, and really isn't going to work
well as we look forward to avx512. So let's set up a safe version and
call it setcomp. It is "safe" because it sets the bit you want, and
doesn't claim you can get a reference to an underlying integer. That
gives us freedom to implement it with an underlying bitfield, among
other helpful things.