SIMD batched texture APIs
There was previously a set of TextureSystem API calls (with multi-point
TextureOptions struct and a fancy VaryingRef helper class) that was a
multi-point interface, but we never did much with it and in retrospect,
it does not use a data organization that is good for future-looking
architectures. We are therefore depcrecating it, it will remain for 1.8
in case anybody is using it, but will thereafter be removed.

But we are adding a new set of batched routines that are organized
differently, we think in a way that will be well-suited for better
parallelism.

The new API assumes that the inputs and outputs have a SOA rather than
AOS layout, and the arrays are a fixed width, defined by
Tex::BatchWidth, which is a constant and set to 16. Build-time overrides
let us change this for experimentation and future expansion. You can
texture fewer points than this using the RunMask, which is a bit field
saying which "lanes" should be used. Inputs will not be read, or outputs
written, for the lanes where the run mask is 0.

The goal is for this to really fly on chips with ISA that includes
AVX-512 support (naturally 16 wide), and to also perform very well for
AVX2 class machines (emulating 16x operations with pairs of 8x SIMD
instructions). I'm not very hopeful that old SSE hardware (4x wide) will
have a high payoff using the batched functionality rather than the
single-point texturing, but we'll see what happens. We'll also
experiment with 32 wide (there's no true 32 wide hardware, but pairs of
16-wide instructions may dual-issue in nice ways that provides an
advantage, we'll see).

I'm not planning to ever remove the single-point scalar API calls. There
will always be a use for them and they'll certainly be more efficient
for applications that don't have a good way to batch the calls.

This pull request only implements the batched APIs -- the internals are
still scalar, they just loop over the enabled lanes and call the scalar
functionality.

I'm actually working on the true batched reversions right now, but they
aren't ready to review yet. Expect a second PR for that as soon as I
have good results. But in the mean time, this allows applications to
start experimenting with the batched interface.

The testtex program now has modes that exercise the batched interfaces,
and the testsuite has been doctored so that every test whose name
contains the substring "texture" will run twice: once using the old
scalar mode, and once using the batched API calls. This means that
starting now, we will immediately notice any breakages or regressions in
the batched versions, as well as have a convenient way to benchmark the
scalar versus batched calls.