strided_ptr<> improvements
Templatize based on the stride units. The default strided_ptr<T>
is equivalent to strided_ptr<T,sizeof(T)>, which means that strides
are measured in terms of the size of the element type. But, if you
really want to, you could make strided_ptr<T,1> which would be a
strided pointer where the strides are bytes rather than T's.

Note that this is a compatibility break -- previously, strided_ptr<>
used byte addressing all the time. This was counter-intuitive in many
circumstances, so now the default is to use array-like addressing where
the strides are measured in units of sizeof(T)>. But we have the ability
to still use byte addressing when really needed.

While I was in there, I changed a lot of the declarations to use
constexpr and noexcept very liberally, to maximize the chances that the
compiler can exploit certain optimizations.