Fix important performance issues with IBA::deep_merge()
When merging, high number of splits was crippling performance due to
constant realocation.

New strategy more carefully pre-computes the NUMBER of splits expected
in the merged pixel, and allocates enough storage for the result pixel
to hold the number of samples in both of the input images, and also the
expected number of new segments that will result from splitting.

This speeds things up for very complicated frames by several orders of
magnitude.