Simplify array_view and array_view_strided to 1D only. (1734)
These were templated on dimensionality of the array, and that made the
implementation much more complex and quite hard to read and
understand. But we never used anything but the 1D case. The reason it
was templated on rank before was to emulate the array_view that was
proposed for C++17, but that has since been scrapped, so keeping
compatibility with it is no longer a priority.

When we remove "Rank" from the template and take away all the machinery
that isn't needed for the simple 1D case, a lot of the complexity goes
away, and these templates will be a lot easier to understand and modify
in the future.