Bump minimum OpenEXR/Ilmbase to 2.0
VFXPlatform has dictated a minimum OpenEXR version of 2.0 since 2014,
and 2.2 since 2015.

Upgrading our minimum to 2.0 eliminates a lot of #if nonsense that was
necessary to continue supporting OpenEXR 1.x. There's no real burden
in continuing to support 2.0 & 2.1 for now. But we will continue to
re-evaluate the appropriate minimum in future releases.