Fix statics contributing to static initialization order fiasco
Fixes #1750

These two spots were implicated in the initialization order fiasco.
(See: https://isocpp.org/wiki/faq/ctors#static-init-order )

The solution for maketx.cpp is that since we only are using the
ColorConfig for the sake of generating a help message, there's no reason
it can't be an ordinary local variable in the one (rarely called)
function where it's used.

For the std::locale involved in strutil.cpp, use the old trick where
instead of a file static, we have a function returning a reference,
which itself has a function-static pointer that's allocated the first
time it's needed. This leaks (a single std::locale), but who cares?