Protect against bogus texture SHA-1 and other attributes (#1762)
It came to my attention that somebody could take a maketx-generated (or,
`oiiotool -otex`, same thing) texture, load it in Photoshop or other
utility, change it, and save it, and it just might still have some of
the maketx-specific metadata in the altered file, such as the SHA-1 hash
that allows us to find duplicates, and those data may be totally wrong
in the new file.

So this patch is intended to do a sanity check: if there is no
"textureformat" metadata present, or if it's not tiled, or if the
"Software" metadata doesn't say maketx or OpenImageIO, then it can't
have come straight out of maketx/oiiotool, so erase any of the
problematic attributes that maketx writes, because they are very likely
wrong and no longer match the potentially modified pixel values in the
file.

Photoshop always writes scanline images, so the check for tiling works
for that case. It will catch others as well. Not perfect by any means,
but it's better than before and I'm not sure if we can ever be certain.
We still rely on users not being foolish -- please only use textures
that have passed through `maketx` or `oiiotool -otex` and don't modify
the pixels after that step without re-maketxing it.