TypeDesc: add constexpr typedescs at the OIIO scope (#1761)
Previously, we had pre-constructed TypeDescs of common types as static
data members of TypeDesc itself. This presented two (minor) problems:
1. Because their full definitions were in typedesc.cpp, all other modules
did not know what was actually in the fields, so optimization opportunities
were potentially missed.
2. There existed the potential for the static initialization order fiasco,
if the compiler could not figure out that the definitions in typedesc.cpp
could be expressed as POD and baked into the module without running the
constructor at module load time.

So we are deprecating these, though they will stick around for another
major release in order to not break any apps that are depending on them.

In their place, we are making OIIO-scoped 'static constexpr' versions.
These serve the same purpose, but by being constexpr (yay, C++11!),
they are should be turned into compile-time constants everyplace they
are used, which is what we ideally wanted all along.