oiiotool: Revamp of logic for propagating data format.
So, oiiotool reads one or more images, might combine them, might
generate some from whole cloth. And the ones read from disk might not be
stored internally with the same data format as it existed on disk (it's
biased to store in float internally). How does it decide what data
format to write to disk for output?

We thought we were doing something sensible -- which is to allow an
explicit override (via `-d`), but in other cases to write using the
same data format as whatever was the FIRST file input. That at least
guarantees that a basic read-write cycle preserved the format.

But Chris Allen pointed out an interesting corner case, which is that
when you use `--siappend`, it will write out all subimages using the
first format it encountered.  But it seems like in this case, it ought
to be able to keep each subimage in the format it started in.

This sent me down a bit of a rabbit hole.

The logic is hereby modified to:

1. Explicit format requests (`-d`) are honored.
2. Otherwise, a file that appears to be an unmodified copy of a file read
   from disk will write using its native (i.e., from disk) format.
3. Otherwise (for a modified or generated file), it will write using
   a data format taken from whatever was the first file read.
4. If all those fail, it will just use the format of the buffer.

Note that (2) is the new bit of logic.

Turned out to take a fair bit of refactoring to really track this
through all the corner cases, and also make it understand that the
"subimage append" operation does not lose the "it's a directly read
image" property.

I implemented similar logic for tracking tile sizes as well. So output
tiling should more accurately reflect input tiling, except when overridden.

Some other miscellaneous things that came along for the ride:

* New testsuite entry: oiiotool-copy to test these cases.
* A bit more care internal to ImageBuf::copy() of keeping track of the
  "native" format info.
* Some C++ modernization of some ImageRec and Oiiotool internals.
* Minor fixes to how regex matching is done for `oiiotool -info` when
  `-metamatch` is used.
* The `-o` command now takes additional arguments: `scanline=1`,
  `tile=XxY`, which can override the tile size or scanlineness
  on a per-output basis (just like how it could override data type
  or contig/separate).