testtex: make the tex_thread_workout cases all honor --handle
There was a --handle flag that lets you test performance differences
between the texture calls that take handles, versus the ones that take
filenames and figure out the handles by looking it up in the file
cache. But it turns out I had implemented that option fully for the
usual "make a texture image" tests, but not for the special thread
workload test you get with the --threadtimes option.