Refactor of metadata printing and Exif read/write (#1774)
"Maker notes" are extra camera manufacturer-specific metadata stored as
opaque blobs inside Exif records. The manufacturers don't seem to
publish the specs for these blobs, but they have been reverse engineered
and with some hunting you can find tinkerer-prouced documentation (at a
range of levels of completeness or correctness).

With this patch, we do a substantial refactor of exif.cpp (and a bit of
work in formatspec, rawinput, and other places) to incorporate full
awareness of maker notes into our metadata processing. At this stage, we
only read Canon makernotes, and quite incompletely. But we get a lot
more of the metadata than we used to (previously we just ignored any
makernotes we found), and the structure is now sound for adding more
complete support for this and other cameras.

This will be an ongoing project to support more cameras, but the foundation
is now set for this work.